Fix search filter and handle validation errors

+++ src/main/java/com/abn/recipe/service/RecipeService.java
@@ -9,7 +9,6 @@ import com.abn.recipe.repository.RecipeEntity_;
 import com.abn.recipe.repository.RecipeRepository;
 import jakarta.persistence.criteria.Predicate;
 import jakarta.transaction.Transactional;
-import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.domain.Specification;
 import org.springframework.stereotype.Service;
@@ -30,9 +29,9 @@ public class RecipeService {
     }
 
     @Transactional
-    public void saveRecipe(Recipe recipe) {
+    public long saveRecipe(Recipe recipe) {
         RecipeEntity recipeEntity = recipeAdapter.toEntity(recipe);
-        recipeRepository.save(recipeEntity);
+        return recipeRepository.save(recipeEntity).getId();
     }
 
     public Recipe getRecipeById(Long id) {
@@ -55,11 +54,14 @@ public class RecipeService {
 
     @Transactional
     public void deleteRecipe(Long id) {
+        if (!recipeRepository.existsById(id)) {
+            throw new RecipeNotFoundException(id);
+        }
         recipeRepository.deleteById(id);
     }
 
     public List<Recipe> getAllRecipes(Pageable pageable) {
-        return recipeRepository.findAll(pageable).stream().map(recipeAdapter::toDomain).collect(Collectors.toList());
+        return recipeRepository.findAll(pageable).stream().map(recipeAdapter::toDomain).toList();
     }
 
     public List<Recipe> searchRecipes(RecipeSearch recipeSearch, Pageable pageable) {
@@ -72,6 +74,10 @@ public class RecipeService {
             RecipeSearch recipeSearch) {
         return (root, query, criteriaBuilder) -> {
             List<Predicate> predicates = new ArrayList<>();
+
+            if (recipeSearch.getName() != null) {
+                predicates.add(criteriaBuilder.equal(root.get(RecipeEntity_.name), recipeSearch.getName()));
+            }
 
             if (recipeSearch.getVegetarian() != null) {
                 predicates.add(criteriaBuilder.equal(root.get(RecipeEntity_.vegetarian), recipeSearch.getVegetarian()));
@@ -91,6 +97,10 @@ public class RecipeService {
                 recipeSearch.getExcludeIngredients().forEach(ingredient ->
                         predicates.add(criteriaBuilder.isNotMember(ingredient, root.get(RecipeEntity_.ingredients)))
                 );
+            }
+
+            if (recipeSearch.getInstructionText() != null && !recipeSearch.getInstructionText().isEmpty()) {
+                predicates.add(criteriaBuilder.like(root.get(RecipeEntity_.instructions), "%" + recipeSearch.getInstructionText() + "%"));
             }
 
             return criteriaBuilder.and(predicates.toArray(new Predicate[0]));
